tweak testing doc

+++ tests/script_doctrine_dbal.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -e
 
 SCRIPT_DIR="${0%/*}"
 
@@ -12,6 +13,8 @@ elif test "${SCRIPT_DIR:0:1}" != "/" ; then
   SCRIPT_DIR="$PWD/$SCRIPT_DIR"
 fi
 
-cp ${SCRIPT_DIR}/../cli-config.doctrine_dbal.php.dist ${SCRIPT_DIR}/../cli-config.php
+if [ ! -f ${SCRIPT_DIR}/../cli-config.php ]; then
+  cp ${SCRIPT_DIR}/../cli-config.doctrine_dbal.php.dist ${SCRIPT_DIR}/../cli-config.php
+fi
 ${SCRIPT_DIR}/../bin/phpcrodm jackalope:init:dbal --force
-  ${SCRIPT_DIR}/../bin/phpcrodm doctrine:phpcr:register-system-node-types
+${SCRIPT_DIR}/../bin/phpcrodm doctrine:phpcr:register-system-node-types
